+++ CHANGELOG.md
@@ -15,6 +15,12 @@
 - выделена задача для установки nuget
 - добавлена задача сборки 'distclean' в целях соответствия
   [стандартным целям make](https://www.gnu.org/software/make/manual/html_node/Standard-Targets.html)
+- добавлена задача сборки 'maintainer-clean' в целях соответствия
+  [стандартным целям make](https://www.gnu.org/software/make/manual/html_node/Standard-Targets.html)
+- добавлена задача сборки 'check' в целях соответствия
+  [стандартным целям make](https://www.gnu.org/software/make/manual/html_node/Standard-Targets.html)
+- добавлена задача сборки 'all' в целях соответствия
+  [стандартным целям make](https://www.gnu.org/software/make/manual/html_node/Standard-Targets.html)
 - переименованы задачи сборки в целях соответствия
   [стандартным целям make](https://www.gnu.org/software/make/manual/html_node/Standard-Targets.html)